+++ tests/test_utils.py
@@ -2,12 +2,12 @@ import wagtail_factories
 from django.test import TestCase
 from graphql import GraphQLError
 
-from grapple.utils import resolve_site
+from grapple.utils import resolve_site_by_hostname, resolve_site_by_id
 
 
-class TestResolveSite(TestCase):
+class TestResolveSiteById(TestCase):
     """
-    Test suite for the `grapple.utils.resolve_site()` utility method.
+    Test suite for the `grapple.utils.resolve_site_by_id()` utility method.
     """
 
     def test_success_with_id(self):
@@ -17,7 +17,27 @@ class TestResolveSite(TestCase):
 
         site = wagtail_factories.SiteFactory(id=1)
 
-        self.assertEqual(resolve_site(id=1), site)
+        self.assertEqual(resolve_site_by_id(id=1), site)
+
+    def test_response_is_none_when_no_sites_match_id(self):
+        """
+        Ensure `None` is returned when passing a `id` that doesn't match
+        any `Site`s.
+        """
+
+        wagtail_factories.SiteFactory()
+
+        self.assertEqual(
+            resolve_site_by_id(id=999999999),
+            None,
+        )
+
+
+class TestResolveSiteByHostname(TestCase):
+    """
+    Test suite for the `grapple.utils.resolve_site_by_hostname()` utility
+    method.
+    """
 
     def test_success_with_ambiguous_hostname(self):
         """
@@ -28,7 +48,10 @@ class TestResolveSite(TestCase):
         site = wagtail_factories.SiteFactory(hostname="example.com")
 
         self.assertEqual(
-            resolve_site(hostname="example.com", hostname_filter_name="hostname"),
+            resolve_site_by_hostname(
+                hostname="example.com",
+                hostname_filter_name="hostname",
+            ),
             site,
         )
 
@@ -41,8 +64,27 @@ class TestResolveSite(TestCase):
         site = wagtail_factories.SiteFactory(hostname="example.com", port="1000")
 
         self.assertEqual(
-            resolve_site(hostname="example.com:1000", hostname_filter_name="hostname"),
+            resolve_site_by_hostname(
+                hostname="example.com:1000",
+                hostname_filter_name="hostname",
+            ),
             site,
+        )
+
+    def test_response_is_none_when_no_sites_match_hostname(self):
+        """
+        Ensure `None` is returned when passing a `hostname` that doesn't match
+        any `Site`s.
+        """
+
+        wagtail_factories.SiteFactory(hostname="example.com")
+
+        self.assertEqual(
+            resolve_site_by_hostname(
+                hostname="not.example.com",
+                hostname_filter_name="hostname",
+            ),
+            None,
         )
 
     def test_graphqlerror_when_hostname_is_ambiguous(self):
@@ -57,44 +99,7 @@ class TestResolveSite(TestCase):
         with self.assertRaisesRegex(
             GraphQLError, "Try including a port number to disambiguate"
         ):
-            resolve_site(hostname="example.com", hostname_filter_name="hostname")
-
-    def test_response_is_none_when_no_sites_match_hostname(self):
-        """
-        Ensure `None` is returned when passing a `hostname` that doesn't match
-        any `Site`s.
-        """
-
-        wagtail_factories.SiteFactory(hostname="example.com")
-
-        self.assertEqual(
-            resolve_site(hostname="not.example.com", hostname_filter_name="hostname"),
-            None,
-        )
-
-    def test_must_provide_id_or_hostname(self):
-        """
-        Ensure resolve_site() warns if used without `id` or `hostname`.
-        """
-
-        with self.assertRaisesRegex(TypeError, "neither were passed"):
-            resolve_site()
-
-    def test_cannot_provide_id_and_hostname(self):
-        """
-        Ensure resolve_site() warns if used with `id` and `hostname`.
-        """
-
-        with self.assertRaisesRegex(ValueError, "both were passed"):
-            resolve_site(
-                id=1000, hostname="example.com", hostname_filter_name="hostname"
+            resolve_site_by_hostname(
+                hostname="example.com",
+                hostname_filter_name="hostname",
             )
-
-    def test_must_provide_hostname_filter_name_if_using_hostname(self):
-        """
-        Ensure resolve_site() warns if `hostname` is used without
-        `hostname_filter_name`.
-        """
-
-        with self.assertRaisesRegex(TypeError, "hostname_filter_name"):
-            resolve_site(hostname="example.com")